requirements.txt
bcrypt==4.1.2
argon2-cffi==23.1.0
click==8.1.7
crc32c==2.4
